Fix a compatibility for input() between vim and nvim
"input()" of Vim doesn't accept a dictionary as it's argument, but the
function of Neovim does. So add a simple wrapper for "input()" to fix
this incompatibility. "inputdialog()" has already been removed in Neovim
and instead it seems like "input()" of Neovim can accept
"inputdialog()"'s arguments as a dictionary.